@@ -12,14 +12,14 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": ">=5.4.0",
+        "php": ">=7.0",
         "league/container": "^2.2",
         "vlucas/phpdotenv": "^2.4",
         "symfony/finder": "^3.1",
         "symfony/console": "~2.8|~3.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.2",
+        "phpunit/phpunit": "^6.0",
         "squizlabs/php_codesniffer": "^2.7",
         "friendsofphp/php-cs-fixer": "^2.0",
         "phpmd/phpmd": "^2.4"
@@ -29,7 +29,12 @@
     },
     "autoload" : {
         "psr-4": {
-            "SideKit\\Config\\": ["./src", "./tests/"]
+            "Da\\Config\\": ["./src"]
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Da\\Config\\Tests\\": "./tests"
         }
     }
 }
