@@ -1,5 +1,15 @@
 <?php
-namespace SideKit\Config\Exception;
+
+/*
+ * This file is part of the 2amigos/yii2-config-kit project.
+ *
+ * (c) 2amigOS! <http://2amigos.us/>
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
+ */
+
+namespace Da\Config\Exception;
 
 use BadFunctionCallException;
 
